Move to a proper project file structure
