+++ steam_analyzer/steamspy/app_details.py
